+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Changes:
 *   Improve attribute inference attack documentation ([#340](https://github.com/AI-SDC/SACRO-ML/pull/340))
+*   Refactor `Target` class - standardise use of constructor to wrap data ([#342](https://github.com/AI-SDC/SACRO-ML/pull/342))
 
 ## Version 1.3.0 (Jun 17, 2025)